fixes issue #11

+++ src/forms/settings-dialog.cpp
@@ -31,7 +31,7 @@ with this program. If not, see <https://www.gnu.org/licenses/>
 #include <qcheckbox.h>
 
 #include "../version.h"
-PluginWindow::PluginWindow(QWidget *parent)
+PluginWindow::PluginWindow(QWidget* parent)
 	: QDialog(parent, Qt::Dialog), ui(new Ui::PluginWindow)
 {
 	ui->setupUi(this);
@@ -54,9 +54,9 @@ PluginWindow::PluginWindow(QWidget *parent)
 		SLOT(on_check_enabled_state_changed(int)));
 	connect(ui->bidirectional, SIGNAL(stateChanged(int)), this,
 		SLOT(on_bid_enabled_state_changed(int)));
-	
+
 	//Connections for Configure Tab
-	
+
 	connect(ui->cb_obs_output_scene, SIGNAL(currentTextChanged(QString)),
 		this, SLOT(get_sources(QString)));
 	connect(ui->cb_obs_output_action, SIGNAL(currentTextChanged(QString)),
@@ -100,7 +100,8 @@ void PluginWindow::ToggleShowHide()
 
 		setVisible(true);
 
-	} else {
+	}
+	else {
 		setVisible(false);
 		ui->btn_Listen_many->setChecked(false);
 		ui->btn_Listen_one->setChecked(false);
@@ -128,7 +129,8 @@ void PluginWindow::SetAvailableDevices()
 		ui->check_enabled->setEnabled(false);
 		this->ui->outbox->setEnabled(false);
 		ui->tabWidget->setEnabled(false);
-	} else if (midiDevices.size() > 0) {
+	}
+	else if (midiDevices.size() > 0) {
 		for (int i = 0; i < midiDevices.size(); i++) {
 			this->ui->list_midi_dev->addItem(midiDevices.at(i));
 		}
@@ -150,10 +152,10 @@ void PluginWindow::SetAvailableDevices()
 		this->ui->outbox->setCurrentIndex(1);
 #elif __linux__
 		this->ui->outbox->setCurrentIndex(0);
-// linux
+		// linux
 #elif __unix__ // all unices not caught above
 		this->ui->outbox->setCurrentIndex(0);
-// Unix
+		// Unix
 #endif
 		if (midiDevices.size() != 0) {
 			desconnect = connect(
@@ -199,13 +201,14 @@ int PluginWindow::on_check_enabled_state_changed(int state)
 				selectedOutDeviceName.c_str());
 		if (device == NULL) {
 			GetDeviceManager()->RegisterMidiDevice(devicePort,
-							       deviceOutPort);
+				deviceOutPort);
 			device = GetDeviceManager()->GetMidiDeviceByName(
 				selectedDeviceName.c_str());
 			device->open_midi_input_port(devicePort);
 			device->open_midi_output_port(deviceOutPort);
 
-		} else {
+		}
+		else {
 
 			device->open_midi_input_port(devicePort);
 			device->open_midi_output_port(deviceOutPort);
@@ -215,10 +218,10 @@ int PluginWindow::on_check_enabled_state_changed(int state)
 		ui->bidirectional->setEnabled(true);
 		ui->bidirectional->setChecked(true);
 		ui->outbox->setEnabled(true);
-		
+
 	}
 	//ui->outbox->setCurrentText(QString::fromStdString(device->GetOutName()));
-	
+
 	GetConfig()->Save();
 	//on_device_select(ui->list_midi_dev->currentItem()->text());
 	return state;
@@ -228,7 +231,7 @@ void PluginWindow::on_device_select(QString curitem)
 {
 	auto devicemanager = GetDeviceManager();
 	auto config = GetConfig();
-	MidiAgent *MAdevice = devicemanager->GetMidiDeviceByName(curitem);
+	MidiAgent* MAdevice = devicemanager->GetMidiDeviceByName(curitem);
 
 	ui->tabWidget->setTabText(1, QString("Configure - ").append(curitem));
 	// Pull info on if device is enabled, if so set true if not set false
@@ -239,7 +242,8 @@ void PluginWindow::on_device_select(QString curitem)
 		ui->bidirectional->setChecked(MAdevice->isBidirectional());
 		ui->outbox->setCurrentText(MAdevice->get_midi_output_name());
 
-	} else {
+	}
+	else {
 		ui->check_enabled->setChecked(false);
 		ui->outbox->setEnabled(false);
 		ui->bidirectional->setEnabled(false);
@@ -256,24 +260,25 @@ void PluginWindow::handle_midi_message(MidiMessage mess)
 {
 	if (ui->tabWidget->currentIndex() == 1) {
 		if (ui->btn_Listen_one->isChecked() ||
-		    ui->btn_Listen_many->isChecked()) {
+			ui->btn_Listen_many->isChecked()) {
 			blog(1,
-			     "got midi message via gui, \n Device = %s \nMType = %s \n NORC : %i \n Channel: %i \n Value: %i",
-			     mess.device_name.toStdString().c_str(),
-			     mess.message_type.toStdString().c_str(), mess.NORC,
-			     mess.channel, mess.value);
+				"got midi message via gui, \n Device = %s \nMType = %s \n NORC : %i \n Channel: %i \n Value: %i",
+				mess.device_name.toStdString().c_str(),
+				mess.message_type.toStdString().c_str(), mess.NORC,
+				mess.channel, mess.value);
 			ui->mapping_lbl_device_name->setText(mess.device_name);
 			ui->sb_channel->setValue(mess.channel);
 			ui->sb_norc->setValue(mess.NORC);
 			ui->slider_value->setValue(mess.value);
 			ui->cb_mtype->setCurrentText(mess.message_type);
 			ui->btn_Listen_one->setChecked(false);
-		} else {
+		}
+		else {
 			if (map_location(mess) != -1) {
 				ui->table_mapping->selectRow(
 					map_location(mess));
 			}
-			
+
 		}
 	}
 }
@@ -285,11 +290,12 @@ int PluginWindow::on_bid_enabled_state_changed(int state)
 		device->setBidirectional(state);
 		return 1;
 
-	} else {
+	}
+	else {
 		device->setBidirectional(state);
 		return 0;
 	}
-	
+
 }
 
 PluginWindow::~PluginWindow()
@@ -316,12 +322,12 @@ void PluginWindow::get_scene_names()
 void PluginWindow::add_midi_device(QString name)
 {
 	blog(LOG_DEBUG, "Adding Midi Device %s", name.toStdString().c_str());
-	QTableWidgetItem *device_name = new QTableWidgetItem();
-	QTableWidgetItem *device_enabled = new QTableWidgetItem();
-	QTableWidgetItem *device_status = new QTableWidgetItem();
-	QTableWidgetItem *feedback_enabled = new QTableWidgetItem();
-	QTableWidgetItem *feedback_name = new QTableWidgetItem();
-	QTableWidgetItem *feedback_status = new QTableWidgetItem();
+	QTableWidgetItem* device_name = new QTableWidgetItem();
+	QTableWidgetItem* device_enabled = new QTableWidgetItem();
+	QTableWidgetItem* device_status = new QTableWidgetItem();
+	QTableWidgetItem* feedback_enabled = new QTableWidgetItem();
+	QTableWidgetItem* feedback_name = new QTableWidgetItem();
+	QTableWidgetItem* feedback_status = new QTableWidgetItem();
 	int rowcount = this->ui->table_mapping->rowCount();
 	this->ui->table_mapping->insertRow(rowcount);
 	device_name->setText(name);
@@ -341,9 +347,9 @@ void PluginWindow::add_midi_device(QString name)
 void PluginWindow::set_headers()
 {
 	ui->table_mapping->setHorizontalHeaderLabels(
-		{"Channel", "Message Type", "Note or Control", "Action",
+		{ "Channel", "Message Type", "Note or Control", "Action",
 		 "Scene", "Source", "Filter", "Transition", "Item",
-		 "Audio Source", "Media Source"});
+		 "Audio Source", "Media Source" });
 	QColor midicolor("#00aaff");
 
 	QColor actioncolor("#aa00ff");
@@ -529,15 +535,15 @@ void PluginWindow::ShowOnly(QList<ActionsClass::Actions> shows)
 	for (int i = 0; i < shows.size(); i++) {
 		ui->cb_obs_output_action->addItem(
 			obs_module_text(ActionsClass::action_to_string(shows.at(i))
-						.toStdString()
-						.c_str()));
+				.toStdString()
+				.c_str()));
 	}
 }
 
 void PluginWindow::ShowEntry(ActionsClass::Actions Entry)
 {
 	if (ui->cb_obs_output_action->findText(
-		    ActionsClass::action_to_string(Entry)) == -1) {
+		ActionsClass::action_to_string(Entry)) == -1) {
 		ui->cb_obs_output_action->addItem(obs_module_text(
 			ActionsClass::action_to_string(Entry).toStdString().c_str()));
 	}
@@ -545,7 +551,7 @@ void PluginWindow::ShowEntry(ActionsClass::Actions Entry)
 void PluginWindow::HideEntry(ActionsClass::Actions Entry)
 {
 	if (ui->cb_obs_output_action->findText(
-		    ActionsClass::action_to_string(Entry)) > 0) {
+		ActionsClass::action_to_string(Entry)) > 0) {
 		ui->cb_obs_output_action->removeItem(
 			ui->cb_obs_output_action->findText(
 				ActionsClass::action_to_string(Entry)));
@@ -589,7 +595,7 @@ void PluginWindow::get_sources(QString scene)
 	auto arrayref = Utils::GetSceneArray(scene);
 	int size = obs_data_array_count(arrayref);
 	for (int i = 0; i < size; i++) {
-		obs_data *item = obs_data_array_item(arrayref, i);
+		obs_data* item = obs_data_array_item(arrayref, i);
 
 		ui->cb_obs_output_source->addItem(
 			QString(obs_data_get_string(item, "name")));
@@ -600,7 +606,7 @@ void PluginWindow::get_sources(QString scene)
 void PluginWindow::get_scenes()
 {
 	ui->cb_obs_output_scene->clear();
-	obs_data_array *x = Utils::GetScenes();
+	obs_data_array* x = Utils::GetScenes();
 	int cnt = obs_data_array_count(x);
 	for (int i = 0; i <= cnt; i++) {
 		auto it = obs_data_array_item(x, i);
@@ -745,9 +751,9 @@ bool PluginWindow::map_exists()
 		ui->mapping_lbl_device_name->text());
 	for (int i = 0; i < hooks.size(); i++) {
 		if ((hooks.at(i)->channel == ui->sb_channel->value()) &&
-		    (hooks.at(i)->norc == ui->sb_norc->value()) &&
-		    (hooks.at(i)->message_type ==
-		     ui->cb_mtype->currentText())) {
+			(hooks.at(i)->norc == ui->sb_norc->value()) &&
+			(hooks.at(i)->message_type ==
+				ui->cb_mtype->currentText())) {
 			return true;
 		}
 	}
@@ -760,9 +766,9 @@ int PluginWindow::map_location(MidiMessage message)
 		ui->mapping_lbl_device_name->text());
 	for (int i = 0; i < hooks.size(); i++) {
 		if ((hooks.at(i)->channel == message.channel) &&
-		    (hooks.at(i)->norc == message.NORC) &&
-		    (hooks.at(i)->message_type ==
-		     message.message_type)) {
+			(hooks.at(i)->norc == message.NORC) &&
+			(hooks.at(i)->message_type ==
+				message.message_type)) {
 			return i;
 		}
 	}
@@ -772,31 +778,31 @@ void PluginWindow::add_new_mapping()
 {
 	ui->btn_Listen_many->setChecked(false);
 	ui->btn_Listen_one->setChecked(false);
-	if (!map_exists()) {
+	if (!map_exists() && verify_mapping()) {
 
 		int row = ui->table_mapping->rowCount();
 		ui->table_mapping->insertRow(row);
-		QTableWidgetItem *channelitem = new QTableWidgetItem(
+		QTableWidgetItem* channelitem = new QTableWidgetItem(
 			QString::number(ui->sb_channel->value()));
-		QTableWidgetItem *mtypeitem =
+		QTableWidgetItem* mtypeitem =
 			new QTableWidgetItem(ui->cb_mtype->currentText());
-		QTableWidgetItem *norcitem = new QTableWidgetItem(
+		QTableWidgetItem* norcitem = new QTableWidgetItem(
 			QString::number(ui->sb_norc->value()));
-		QTableWidgetItem *actionitem = new QTableWidgetItem(
+		QTableWidgetItem* actionitem = new QTableWidgetItem(
 			ui->cb_obs_output_action->currentText());
-		QTableWidgetItem *sceneitem = new QTableWidgetItem(
+		QTableWidgetItem* sceneitem = new QTableWidgetItem(
 			ui->cb_obs_output_scene->currentText());
-		QTableWidgetItem *sourceitem = new QTableWidgetItem(
+		QTableWidgetItem* sourceitem = new QTableWidgetItem(
 			ui->cb_obs_output_source->currentText());
-		QTableWidgetItem *filteritem = new QTableWidgetItem(
+		QTableWidgetItem* filteritem = new QTableWidgetItem(
 			ui->cb_obs_output_filter->currentText());
-		QTableWidgetItem *transitionitem = new QTableWidgetItem(
+		QTableWidgetItem* transitionitem = new QTableWidgetItem(
 			ui->cb_obs_output_transition->currentText());
-		QTableWidgetItem *itemitem = new QTableWidgetItem(
+		QTableWidgetItem* itemitem = new QTableWidgetItem(
 			ui->cb_obs_output_item->currentText());
-		QTableWidgetItem *audioitem = new QTableWidgetItem(
+		QTableWidgetItem* audioitem = new QTableWidgetItem(
 			ui->cb_obs_output_audio_source->currentText());
-		QTableWidgetItem *mediaitem = new QTableWidgetItem(
+		QTableWidgetItem* mediaitem = new QTableWidgetItem(
 			ui->cb_obs_output_media_source->currentText());
 		ui->table_mapping->setItem(row, 0, channelitem);
 		ui->table_mapping->setItem(row, 1, mtypeitem);
@@ -809,7 +815,7 @@ void PluginWindow::add_new_mapping()
 		ui->table_mapping->setItem(row, 8, itemitem);
 		ui->table_mapping->setItem(row, 9, audioitem);
 		ui->table_mapping->setItem(row, 10, mediaitem);
-		MidiHook *newmh = new MidiHook();
+		MidiHook* newmh = new MidiHook();
 		newmh->channel = ui->sb_channel->value();
 		newmh->message_type = ui->cb_mtype->currentText();
 		newmh->norc = ui->sb_norc->value();
@@ -829,22 +835,29 @@ void PluginWindow::add_new_mapping()
 		dev->add_MidiHook(newmh);
 		auto conf = GetConfig();
 		conf->Save();
-	} else {
-		// TODO: Fix this, and create Utils::message_to_user(QString)
-		QString Mess;
-		Mess.append("Mapping already Exists for ");
-		Mess.append(ui->mapping_lbl_device_name->text());
-		Mess.append(" , with channel # ");
-		Mess.append(QString::number(ui->sb_channel->value()));
-		Mess.append(" norc # ");
-		Mess.append(QString::number(ui->sb_norc->value()));
-		Mess.append(" and Message Type ");
-		Mess.append(ui->cb_mtype->currentText());
-		Utils::alert_popup(Mess);
+	}
+	else {
+		if (!verify_mapping()) {
+			Utils::alert_popup("Mapping Missing required variable");
+		}
+		if(map_exists()){
+			// TODO: Fix this, and create Utils::message_to_user(QString)
+			QString Mess;
+
+			Mess.append("Mapping already Exists for ");
+			Mess.append(ui->mapping_lbl_device_name->text());
+			Mess.append(" , with channel # ");
+			Mess.append(QString::number(ui->sb_channel->value()));
+			Mess.append(" norc # ");
+			Mess.append(QString::number(ui->sb_norc->value()));
+			Mess.append(" and Message Type ");
+			Mess.append(ui->cb_mtype->currentText());
+			Utils::alert_popup(Mess);
+		}
 	}
 }
 
-void PluginWindow::add_row_from_hook(MidiHook *hook)
+void PluginWindow::add_row_from_hook(MidiHook* hook)
 {
 	int row = ui->table_mapping->rowCount();
 	ui->table_mapping->insertRow(row);
@@ -852,21 +865,21 @@ void PluginWindow::add_row_from_hook(MidiHook *hook)
 	midic.setRgb(0, 170, 255);
 	QColor actc;
 	actc.setRgb(170, 0, 255);
-	QTableWidgetItem *channelitem =
+	QTableWidgetItem* channelitem =
 		new QTableWidgetItem(QString::number(hook->channel));
 
-	QTableWidgetItem *mtypeitem = new QTableWidgetItem(hook->message_type);
-	QTableWidgetItem *norcitem =
+	QTableWidgetItem* mtypeitem = new QTableWidgetItem(hook->message_type);
+	QTableWidgetItem* norcitem =
 		new QTableWidgetItem(QString::number(hook->norc));
-	QTableWidgetItem *actionitem = new QTableWidgetItem(hook->action);
-	QTableWidgetItem *sceneitem = new QTableWidgetItem(hook->scene);
-	QTableWidgetItem *sourceitem = new QTableWidgetItem(hook->source);
-	QTableWidgetItem *filteritem = new QTableWidgetItem(hook->filter);
-	QTableWidgetItem *transitionitem =
+	QTableWidgetItem* actionitem = new QTableWidgetItem(hook->action);
+	QTableWidgetItem* sceneitem = new QTableWidgetItem(hook->scene);
+	QTableWidgetItem* sourceitem = new QTableWidgetItem(hook->source);
+	QTableWidgetItem* filteritem = new QTableWidgetItem(hook->filter);
+	QTableWidgetItem* transitionitem =
 		new QTableWidgetItem(hook->transition);
-	QTableWidgetItem *itemitem = new QTableWidgetItem(hook->item);
-	QTableWidgetItem *audioitem = new QTableWidgetItem(hook->audio_source);
-	QTableWidgetItem *mediaitem = new QTableWidgetItem(hook->media_source);
+	QTableWidgetItem* itemitem = new QTableWidgetItem(hook->item);
+	QTableWidgetItem* audioitem = new QTableWidgetItem(hook->audio_source);
+	QTableWidgetItem* mediaitem = new QTableWidgetItem(hook->media_source);
 
 	set_cell_colors(midic, channelitem);
 	set_cell_colors(midic, mtypeitem);
@@ -892,7 +905,7 @@ void PluginWindow::add_row_from_hook(MidiHook *hook)
 	ui->table_mapping->setItem(row, 9, audioitem);
 	ui->table_mapping->setItem(row, 10, mediaitem);
 }
-void PluginWindow::set_cell_colors(QColor color, QTableWidgetItem *item)
+void PluginWindow::set_cell_colors(QColor color, QTableWidgetItem* item)
 {
 	QColor txcolor;
 	txcolor.black();
@@ -908,7 +921,7 @@ void PluginWindow::tab_changed(int i)
 	set_headers();
 	ui->table_mapping->setRowCount(0);
 	auto devicemanager = GetDeviceManager();
-	
+
 	auto hooks = devicemanager->GetMidiHooksByDeviceName(
 		ui->mapping_lbl_device_name->text());
 	if (hooks.count() > 0) {
@@ -916,7 +929,7 @@ void PluginWindow::tab_changed(int i)
 			add_row_from_hook(hooks.at(i));
 		}
 	}
-	
+
 }
 void PluginWindow::delete_mapping() {
 	auto devicemanager = GetDeviceManager();
@@ -926,18 +939,18 @@ void PluginWindow::delete_mapping() {
 	auto conf = GetConfig();
 	for (int i = 0; i < hooks.size(); i++) {
 		if ((hooks.at(i)->channel == ui->sb_channel->value()) &&
-		    (hooks.at(i)->norc == ui->sb_norc->value()) &&
-		    (hooks.at(i)->message_type ==
-		     ui->cb_mtype->currentText())) {
+			(hooks.at(i)->norc == ui->sb_norc->value()) &&
+			(hooks.at(i)->message_type ==
+				ui->cb_mtype->currentText())) {
 			dev->remove_MidiHook(hooks.at(i));
 			conf->Save();
 			ui->table_mapping->removeRow(
 				ui->table_mapping->selectedItems().at(0)->row());
 		}
 	}
-	
-	
-	
+
+
+
 }
 void PluginWindow::edit_mapping(int row, int col)
 {
@@ -955,4 +968,35 @@ void PluginWindow::edit_mapping(int row, int col)
 	ui->cb_obs_output_item->setCurrentText(sitems.at(8)->text());
 	ui->cb_obs_output_audio_source->setCurrentText(sitems.at(9)->text());
 	ui->cb_obs_output_media_source->setCurrentText(sitems.at(10)->text());
+}
+bool PluginWindow::verify_mapping() {
+	int testresults = 0;
+	if (ui->cb_obs_output_audio_source->isVisible() && ui->cb_obs_output_audio_source->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_scene->isVisible() && ui->cb_obs_output_scene->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_source->isVisible() && ui->cb_obs_output_source->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_filter->isVisible() && ui->cb_obs_output_filter->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_transition->isVisible() && ui->cb_obs_output_transition->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_item->isVisible() && ui->cb_obs_output_item->count() == 0) {
+		testresults++;
+	}
+	if (ui->cb_obs_output_media_source->isVisible() && ui->cb_obs_output_media_source->count() == 0) {
+		testresults++;
+	}
+	
+	if (testresults > 0) {
+		return false;
+	}
+	else {
+		return true;
+	}
 }
